Move normalizeWatchers into its own module
Extract normalizeWatchers from set-value.ts into a dedicated src/runtime/normalize-watchers.ts and remove the duplicate implementation. Update imports across the runtime and hydrate code (bootstrap-custom-element, bootstrap-lazy, initialize-component, proxy-component, runtime/index, hydrate/platform/proxy-host-element) and test to use the new module. The change centralizes watcher normalization (handling legacy string[] watcher format vs. the new {method: flags}[] format) and includes a comment in bootstrap-lazy explaining the compatibility rationale.
