+++ src/runtime/bootstrap-custom-element.ts
@@ -25,7 +25,7 @@ import {
 import { computeMode } from './mode';
 import { proxyComponent } from './proxy-component';
 import { PROXY_FLAGS } from './runtime-constants';
-import { normalizeWatchers } from './set-value';
+import { normalizeWatchers } from './normalize-watchers';
 import { attachStyles, getScopeId, hydrateScopedToShadow, registerStyle } from './styles';
 
 export const defineCustomElement = (Cstr: any, compactMeta: d.ComponentRuntimeMetaCompact) => {
