+++ src/runtime/initialize-component.ts
@@ -8,7 +8,7 @@ import { computeMode } from './mode';
 import { createTime, uniqueTime } from './profile';
 import { proxyComponent } from './proxy-component';
 import { PROXY_FLAGS } from './runtime-constants';
-import { normalizeWatchers } from './set-value';
+import { normalizeWatchers } from './normalize-watchers';
 import { getScopeId, registerStyle } from './styles';
 import { safeCall, scheduleUpdate } from './update-component';
